requirements_web.txt
fastapi
uvicorn[standard]
openai
psycopg2-binary
